♻️ refactor(auth): move auth constants from packages/const to src/envs/auth
- Move enableClerk, enableBetterAuth, enableNextAuth, enableAuth to src/envs/auth.ts
- Move LOBE_CHAT_AUTH_HEADER, LOBE_CHAT_OIDC_AUTH_HEADER, OAUTH_AUTHORIZED, SECRET_XOR_KEY
- Delete packages/const/src/auth.ts
- Update all imports from @/const/auth and @lobechat/const to @/envs/auth
- Update test file mock paths
