--- packages/const/src/auth.ts
@@ -1,14 +1,0 @@
-export const enableClerk =
-  process.env.NEXT_PUBLIC_ENABLE_CLERK_AUTH === '1'
-    ? true
-    : !!process.env.NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY;
-export const enableBetterAuth = process.env.NEXT_PUBLIC_ENABLE_BETTER_AUTH === '1';
-export const enableNextAuth = process.env.NEXT_PUBLIC_ENABLE_NEXT_AUTH === '1';
-export const enableAuth = enableClerk || enableBetterAuth || enableNextAuth || false;
-
-export const LOBE_CHAT_AUTH_HEADER = 'X-lobe-chat-auth';
-export const LOBE_CHAT_OIDC_AUTH_HEADER = 'Oidc-Auth';
-
-export const OAUTH_AUTHORIZED = 'X-oauth-authorized';
-
-export const SECRET_XOR_KEY = 'LobeHub · LobeHub';
