+++ packages/const/src/index.ts
@@ -1,4 +1,3 @@
-export * from './auth';
 export * from './currency';
 export * from './desktop';
 export * from './discover';
