+++ src/app/[variants]/(main)/settings/profile/features/AvatarRow.tsx
@@ -7,7 +7,7 @@ import { useCallback, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { fetchErrorNotification } from '@/components/Error/fetchErrorNotification';
-import { enableAuth } from '@/const/auth';
+import { enableAuth } from '@/envs/auth';
 import UserAvatar from '@/features/User/UserAvatar';
 import { useUserStore } from '@/store/user';
 import { authSelectors } from '@/store/user/selectors';
