+++ src/app/[variants]/(mobile)/me/(home)/features/UserBanner.tsx
@@ -4,7 +4,7 @@ import { Flexbox } from '@lobehub/ui';
 import { memo } from 'react';
 import { Link } from 'react-router-dom';
 
-import { enableAuth } from '@/const/auth';
+import { enableAuth } from '@/envs/auth';
 import DataStatistics from '@/features/User/DataStatistics';
 import UserInfo from '@/features/User/UserInfo';
 import UserLoginOrSignup from '@/features/User/UserLoginOrSignup/Community';
