+++ src/features/User/__tests__/PanelContent.test.tsx
@@ -74,7 +74,7 @@ const { enableAuth, enableClerk, enableNextAuth } = vi.hoisted(() => ({
   enableNextAuth: { value: false },
 }));
 
-vi.mock('@/const/auth', () => ({
+vi.mock('@/envs/auth', () => ({
   get enableAuth() {
     return enableAuth.value;
   },
