+++ src/features/User/__tests__/UserAvatar.test.tsx
@@ -16,7 +16,7 @@ const { enableAuth, enableClerk, enableNextAuth } = vi.hoisted(() => ({
   enableNextAuth: { value: false },
 }));
 
-vi.mock('@/const/auth', () => ({
+vi.mock('@/envs/auth', () => ({
   get enableAuth() {
     return enableAuth.value;
   },
