+++ src/features/User/__tests__/useMenu.test.tsx
@@ -54,7 +54,7 @@ const { enableAuth, enableClerk } = vi.hoisted(() => ({
   enableClerk: { value: true },
 }));
 
-vi.mock('@/const/auth', () => ({
+vi.mock('@/envs/auth', () => ({
   get enableAuth() {
     return enableAuth.value;
   },
