+++ src/libs/oidc-provider/provider.test.ts
@@ -4,7 +4,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 // Mock dependencies
-vi.mock('@/const/auth', () => ({
+vi.mock('@/envs/auth', () => ({
   enableBetterAuth: false,
   enableClerk: false,
   enableNextAuth: false,
